+++ public/locales/en/eservice.json
@@ -136,9 +136,6 @@
         "doubtsQuestion": "Do you have doubts about the implementation to deliver your e-service?",
         "doubtsLink": "See the guide"
       },
-      "download": {
-        "title": "Download"
-      },
       "voucherVerification": {
         "title": "Voucher verification",
         "description": "Devi verificare un voucher ricevuto da un fruitore? Di seguito i link con le istruzioni necessarie",
