+++ public/locales/it/eservice.json
@@ -136,9 +136,6 @@
         "doubtsQuestion": "Hai dubbi sull’implementazione per erogare il tuo e-service?",
         "doubtsLink": "Consulta la guida"
       },
-      "download": {
-        "title": "Download"
-      },
       "voucherVerification": {
         "title": "Verifica voucher",
         "description": "Devi verificare un voucher ricevuto da un fruitore? Di seguito i link con le istruzioni necessarie",
